# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estuaire Emploie API Documentation",
        "description": "Documentation de l'API REST pour l'application mobile Estuaire Emploie - Plateforme de recherche d'emploi au Cameroun",
        "contact": {
            "email": "contact@estuaire-emploie.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Liste des offres d'emploi publiées",
                "operationId": "63963513839a812f56fe5eddf9fe529a",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrer par catégorie",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "query",
                        "description": "Filtrer par localisation",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "contract_type_id",
                        "in": "query",
                        "description": "Filtrer par type de contrat",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "experience_level",
                        "in": "query",
                        "description": "Filtrer par niveau d'expérience",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "junior",
                                "intermediaire",
                                "senior",
                                "expert"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par mots-clés",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Créer une nouvelle offre d'emploi",
                "description": "Permet à un recruteur de créer une nouvelle offre d'emploi pour son entreprise. L'offre sera en attente de validation par l'administrateur.",
                "operationId": "createJob",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "category_id",
                                    "location_id",
                                    "contract_type_id",
                                    "experience_level"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Développeur Full Stack Senior"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nous recherchons un développeur Full Stack avec expertise Laravel et Vue.js"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "location_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "contract_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "salary_min": {
                                        "type": "number",
                                        "example": 500000,
                                        "nullable": true
                                    },
                                    "salary_max": {
                                        "type": "number",
                                        "example": 800000,
                                        "nullable": true
                                    },
                                    "salary_negotiable": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "experience_level": {
                                        "type": "string",
                                        "enum": [
                                            "junior",
                                            "intermediaire",
                                            "senior",
                                            "expert"
                                        ],
                                        "example": "senior"
                                    },
                                    "requirements": {
                                        "type": "string",
                                        "example": "3+ ans d'expérience en PHP/Laravel",
                                        "nullable": true
                                    },
                                    "benefits": {
                                        "type": "string",
                                        "example": "Assurance santé, primes de performance",
                                        "nullable": true
                                    },
                                    "application_deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Offre créée avec succès. En attente de validation."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé à publier des offres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'êtes pas autorisé à publier des offres"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Détails d'une offre d'emploi",
                "operationId": "46795332805e4a49b36760cdb06df2b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'offre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                }
            }
        },
...
        "/api/recruiter/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Mes offres d'emploi (Recruteur)",
                "description": "Récupère toutes les offres d'emploi publiées par le recruteur connecté pour son entreprise",
                "operationId": "getMyJobs",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "pending",
                                "published",
                                "closed",
                                "expired"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres du recruteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Vous n'êtes pas recruteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'êtes pas recruteur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recruiter/dashboard": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Dashboard recruteur",
                "description": "Récupère les statistiques et données essentielles du dashboard recruteur (stats, offres actives, candidatures récentes)",
                "operationId": "getRecruiterDashboard",
                "responses": {
                    "200": {
                        "description": "Données du dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statistics": {
                                            "properties": {
                                                "total_jobs": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "active_jobs": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "total_applications": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "new_applications": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total_views": {
                                                    "type": "integer",
                                                    "example": 1248
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "active_jobs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "recent_applications": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Vous n'êtes pas recruteur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum - Utiliser le token obtenu lors du login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}